Extract ELU history metrics from the original route
