+++ src/core/packages/metrics/server/src/contracts.ts
@@ -9,6 +9,13 @@
 
 import type { Observable } from 'rxjs';
 import type { OpsMetrics } from './metrics';
+
+interface EluHistory {
+  long: number;
+  medium: number;
+  short: number;
+}
+
 /**
  * APIs to retrieves metrics gathered and exposed by the core platform.
  *
@@ -31,6 +38,8 @@ export interface MetricsServiceSetup {
    * ```
    */
   getOpsMetrics$: () => Observable<OpsMetrics>;
+
+  getEluHistory(): EluHistory;
 }
 /**
  * {@inheritdoc MetricsServiceSetup}
